refactor: move code to userManager
Because userService is deprecated, better to use the core userManager
