+++ EMS/core-bundle/config/controllers.xml
@@ -157,7 +157,7 @@
         <service id="EMS\CoreBundle\Controller\Api\AuthTokenLoginController"/>
 
         <service id="EMS\CoreBundle\Controller\Api\UserController">
-            <argument type="service" id="ems.service.user"/>
+            <argument type="service" id="emsco.manager.user"/>
             <tag name="controller.service_arguments"/>
         </service>
