+++ EMS/core-bundle/src/Service/UserService.php
@@ -10,15 +10,11 @@ use EMS\CommonBundle\Entity\EntityInterface;
 use EMS\CoreBundle\Core\UI\Menu;
 use EMS\CoreBundle\Entity\User;
 use EMS\CoreBundle\Entity\UserInterface;
-use EMS\CoreBundle\Exception\NotFoundException;
-use EMS\CoreBundle\Repository\AuthTokenRepository;
 use EMS\CoreBundle\Repository\SearchRepository;
 use EMS\CoreBundle\Repository\UserRepository;
-use EMS\CoreBundle\Roles;
 use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
-use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class UserService implements EntityServiceInterface
 {
@@ -36,26 +32,8 @@ class UserService implements EntityServiceInterface
         private readonly UserRepository $userRepository,
         private readonly SearchRepository $searchRepository,
         private readonly AuthorizationCheckerInterface $authorizationChecker,
-        private readonly AuthTokenRepository $authTokenRepository,
         private readonly array $securityRoles,
     ) {
-    }
-
-    public function authenticate(string $username, ?string $email): string
-    {
-        if (!$this->isGrantedRole(Roles::ROLE_USER_MANAGEMENT)) {
-            throw new AccessDeniedException();
-        }
-
-        $user = $email
-            ? $this->findUserByEmail($email) ?? $this->findUserByUsername($username)
-            : $this->findUserByUsername($username);
-
-        if (!$user instanceof UserInterface) {
-            throw new NotFoundException('User not found');
-        }
-
-        return $this->authTokenRepository->create($user)->getValue();
     }
 
     public function isSuper(): bool
@@ -88,11 +66,6 @@ class UserService implements EntityServiceInterface
         return $this->userRepository->findOneBy(['email' => $email]);
     }
 
-    public function findUserByUsername(string $username): ?User
-    {
-        return $this->userRepository->findOneBy(['username' => $username]);
-    }
-
     public function updateUser(UserInterface $user): UserInterface
     {
         $em = $this->doctrine->getManager();
@@ -114,11 +87,20 @@ class UserService implements EntityServiceInterface
 
     public function getUser(string $username, bool $detachIt = true): ?UserInterface
     {
-        if (null === $user = $this->findUserByUsername($username)) {
+        $user = $this->userRepository->findOneBy(['username' => $username]);
+
+        if (null === $user) {
             return null;
         }
+        if (!$user instanceof UserInterface) {
+            throw new \RuntimeException(\sprintf('Unknown user object class: %s', $user::class));
+        }
 
-        return !$detachIt ? $user : clone $user;
+        if (!$detachIt) {
+            return $user;
+        }
+
+        return clone $user;
     }
 
     public function getCurrentUser(bool $detach = true): UserInterface
